SistemaCarros/requirements.txt
arabic-reshaper==2.1.3
asgiref==3.4.1
asn1crypto==1.5.1
backports.entry-points-selectable==1.1.1
backports.zoneinfo==0.2.1
cement==2.8.2
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
click==8.1.3
colorama==0.4.3
cryptography==37.0.2
cssselect2==0.6.0
distlib==0.3.4
dj-database-url==0.5.0
Django==3.2.10
django-appconf==1.0.5
django-ckeditor==6.2.0
django-classy-tags==3.0.1
django-compressor==3.1
django-cors-headers==3.11.0
django-dynamic-formsets==0.0.8
django-heroku==0.3.1
django-js-asset==1.2.2
django-mathfilters==1.0.0
django-richtextfield==1.6.1
django-sekizai==0.10.0
django-widget-tweaks==1.4.9
django-xhtml2pdf==0.0.4
djangocms-admin-style==2.0.2
djangorestframework==3.10.3
filelock==3.4.0
future==0.16.0
gunicorn==20.1.0
heroku==0.1.4
html5lib==1.1
idna==3.3
importlib-metadata==4.10.0
jmespath==0.10.0
lxml==4.8.0
oscrypto==1.3.0
pathspec==0.5.9
Pillow==9.0.0
platformdirs==2.4.0
psycopg2==2.9.3
pycparser==2.21
pyHanko==0.13.1
pyhanko-certvalidator==0.19.5
PyPDF3==1.0.6
python-bidi==0.4.2
python-dateutil==1.5
python-decouple==3.5
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==5.4.1
qrcode==7.3.1
rcssmin==1.1.0
reportlab==3.6.5
requests==2.26.0
rjsmin==1.2.0
semantic-version==2.8.5
six==1.14.0
sqlparse==0.4.2
svglib==1.2.1
termcolor==1.1.0
tinycss2==1.1.1
tqdm==4.64.0
typing_extensions==4.0.1
tzdata==2022.1
tzlocal==4.2
uritools==4.0.0
urllib3==1.26.7
virtualenv==20.10.0
wcwidth==0.1.9
webencodings==0.5.1
whitenoise==5.3.0
xhtml2pdf==0.2.7
zipp==3.6.0
